Refactor currentAmount on Tax from int to double, implements #61

+++ wolf-cafe-backend/src/main/java/edu/ncsu/csc326/wolfcafe/controller/AuthController.java
@@ -109,7 +109,7 @@ public class AuthController {
      * @return current tax rate as an integer
      */
     @GetMapping("/tax")
-    public ResponseEntity<Integer> getTaxRate() {
+    public ResponseEntity<Double> getTaxRate() {
     		return ResponseEntity.ok(authService.getTaxRate());
     }
